backend/requirements.txt
fastapi==0.136.0
gunicorn==23.0.0
uvicorn[standard]==0.45.0
python-dotenv==1.2.2
httpx==0.28.1
playwright==1.58.0
supabase==2.24.0
google-genai==1.73.1
langgraph==1.1.9
psycopg[binary]==3.2.12
fpdf2==2.8.3
